Add header logic for burger menu on small screens
On mobile there should be a burger menu for page navigation, now there is a javascript for toggle logic for the burger menu. To be sure that header an footer components are fully loaded there now is a Promise in includeHTML function. Therefore in the header js Promise.all can be used and only after header and footer are loaded the eventListeners are added.
Styling for new header html elements also was added, styling hides burger menu on desktop and desktop navigation on mobile.